+++ src/view/About.tsx
@@ -2,12 +2,10 @@ import { Box, Container } from '@mantine/core';
 
 const About = () => {
   return (
-    <Box p='md'>
-      <Container size='xl'>
-        <h1>About</h1>
-        <p>This is a simple example of a React component that renders HTML.</p>
-      </Container>
-    </Box>
+    <>
+      <h1>About</h1>
+      <p>This is a simple example of a React component that renders HTML.</p>
+    </>
   );
 };
